refactor: Move GPS coordinate utilities to webui/lib for proper architecture (Issue #104)
Refactored GPS coordinate conversion utilities from webui/backend/utils/
to webui/lib/ to follow proper webui architecture where shared utilities
belong in webui/lib rather than being backend-specific.

## Changes

### File Structure
- **Created**: webui/lib/ directory for webui-shared utilities
- **Created**: webui/lib/__init__.py with proper exports
- **Moved**: gps_coordinates.py from webui/backend/utils/ to webui/lib/
- **Updated**: webui/backend/utils/gps_coordinates.py now a stub for backward compatibility

### Import Updates
- **lib/gps_exif_lib.py**: Changed imports from webui.backend.utils.gps_coordinates to webui.lib.gps_coordinates
- **Tests/performance/test_gps_coordinates_performance.py**: Updated to import from webui.lib
- **webui/backend/utils/__init__.py**: Updated re-export to import from webui.lib

### Backward Compatibility
- webui/backend/utils/gps_coordinates.py maintained as re-export stub
- All existing imports continue to work via re-export
- webui/backend/utils/__init__.py continues to export GPS functions

### Documentation Updates
- **CLAUDE.md**: Updated import examples and Key Files Reference
- **docs/GPS_COORDINATE_UTILITIES.md**: Added File Locations section, updated all import examples
- **Tests/unit/test_gps_coordinates.py**: Updated header comment with new location

## Architecture Rationale

Before:
```
webui/backend/utils/gps_coordinates.py  ← Wrong (backend-specific)
lib/gps_exif_lib.py imports from ↑      ← Creates bad dependency
```

After:
```
webui/lib/gps_coordinates.py            ← Correct (webui-shared)
lib/gps_exif_lib.py imports from ↑      ← Clean dependency
webui/backend/ can import from ↑        ← Clean dependency
```

## Testing
- ✅ All 70 GPS coordinate unit tests pass
- ✅ All 16 GPS coordinate performance tests pass
- ✅ All 43 GPS EXIF library tests pass
- ✅ All 2127 unit tests collected successfully
- ✅ No import errors or circular dependencies

Related to Issue #104: GPS Coordinate Utilities Refactoring
