+++ Firmware/webui/backend/utils/__init__.py
@@ -37,8 +37,19 @@ _validate_int_enum = _backend_utils._validate_int_enum
 _validate_exposure_time = _backend_utils._validate_exposure_time
 _validate_noise_reduction_mode = _backend_utils._validate_noise_reduction_mode
 
-# Also export items from this package
-from .gps_coordinates import (
+# Also export items from webui.lib (for backward compatibility)
+# Use relative import to parent package since webui may not be in sys.path
+import sys
+from pathlib import Path
+
+# Add webui directory to path if not already present
+# __file__ is .../webui/backend/utils/__init__.py
+# We need .../webui
+webui_dir = Path(__file__).parent.parent.parent
+if str(webui_dir) not in sys.path:
+    sys.path.insert(0, str(webui_dir))
+
+from webui.lib.gps_coordinates import (
     decimal_to_dms,
     dms_to_decimal,
     format_coordinate_display,
